scraping/requirements.txt
Pillow
requests
imutils
beautifulsoup4
python-magic